requirements.txt
alembic==1.5.6
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-expects-json==1.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
numpy==1.21.2
pandas==1.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
threadpoolctl==2.2.0
Werkzeug==1.0.1
